% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.angle}
\alias{.angle}
\title{Calculate angle between two points}
\usage{
.angle(x, y)
}
\arguments{
\item{x}{numeric vector of length 2}

\item{y}{numeric vector of length 2}
}
\value{
angle in radians
}
\description{
Calculate angle between two points
}
\keyword{internal}
