% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.avgangle}
\alias{.avgangle}
\title{Calculate average angle for three points}
\usage{
.avgangle(x, y)
}
\arguments{
\item{x}{numeric vector of length 3}

\item{y}{numeric vector of length 3}
}
\value{
average angle in radians
}
\description{
Calculate average angle for three points
}
\keyword{internal}
