% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constellation.R
\name{.build_plot}
\alias{.build_plot}
\title{Build the final plot}
\usage{
.build_plot(
  nodes,
  poly.Edges,
  node.label,
  label.size,
  max_size,
  node_trans,
  label_repel
)
}
\description{
Build the final plot
}
\keyword{internal}
