% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constellation_feature.R
\name{.build_plot_feature}
\alias{.build_plot_feature}
\title{Build the final plot for feature coloring}
\usage{
.build_plot_feature(
  nodes,
  poly.Edges,
  node.label,
  label.size,
  max_size,
  node_trans,
  label_repel,
  hull_data = NULL,
  hull_alpha = 0.2,
  legend_title = NULL,
  feature_stat = "avg"
)
}
\description{
Build the final plot for feature coloring
}
\keyword{internal}
