% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_hulls}
\alias{.compute_hulls}
\title{Compute hulls for each group}
\usage{
.compute_hulls(nodes, hull_type = "convex", expand = 0.1)
}
\arguments{
\item{nodes}{data frame with node coordinates and hull_group}

\item{hull_type}{"convex" or "concave"}

\item{expand}{expansion factor}
}
\value{
data frame with hull polygon coordinates
}
\description{
Compute hulls for each group
}
\keyword{internal}
