% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_majority_category}
\alias{.compute_majority_category}
\title{Compute majority category for each cluster}
\usage{
.compute_majority_category(cluster_vec, category_vec, cluster_levels)
}
\arguments{
\item{cluster_vec}{vector of cluster assignments}

\item{category_vec}{vector of category values}

\item{cluster_levels}{unique cluster levels}
}
\value{
named vector of majority categories per cluster
}
\description{
Compute majority category for each cluster
}
\keyword{internal}
