% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constellation.R
\name{.create_line_segments}
\alias{.create_line_segments}
\title{Create line segments}
\usage{
.create_line_segments(knn.cl.lines, nodes, exxageration)
}
\description{
Create line segments
}
\keyword{internal}
