% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_maker.R
\name{.edgeMaker}
\alias{.edgeMaker}
\title{Create a curved edge between two points}
\usage{
.edgeMaker(whichRow, len = 100, line.segments, curved = TRUE)
}
\arguments{
\item{whichRow}{row index in line.segments}

\item{len}{number of points along the edge}

\item{line.segments}{data frame with edge information}

\item{curved}{logical, whether to curve the edge}
}
\value{
data frame with edge coordinates
}
\description{
Create a curved edge between two points
}
\keyword{internal}
