% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.expand_hull}
\alias{.expand_hull}
\title{Expand hull outward from centroid}
\usage{
.expand_hull(hull_pts, expand)
}
\arguments{
\item{hull_pts}{matrix of hull points}

\item{expand}{expansion factor}
}
\value{
expanded hull points
}
\description{
Expand hull outward from centroid
}
\keyword{internal}
