% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_shades}
\alias{.generate_shades}
\title{Generate color shades from a base color}
\usage{
.generate_shades(base_color, n)
}
\arguments{
\item{base_color}{base color}

\item{n}{number of shades to generate}
}
\value{
vector of color shades
}
\description{
Generate color shades from a base color
}
\keyword{internal}
