% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.lighten_color}
\alias{.lighten_color}
\title{Lighten a color}
\usage{
.lighten_color(color, amount = 0.5)
}
\arguments{
\item{color}{input color}

\item{amount}{amount to lighten (0-1)}
}
\value{
lightened color
}
\description{
Lighten a color
}
\keyword{internal}
