% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.perp}
\alias{.perp}
\title{Calculate perpendicular points}
\usage{
.perp(x, y, len, a, mid)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{len}{half-width}

\item{a}{angle}

\item{mid}{midpoint index}
}
\value{
matrix with upper and lower points
}
\description{
Calculate perpendicular points
}
\keyword{internal}
