% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.predict_knn}
\alias{.predict_knn}
\title{Predict cluster membership using KNN}
\usage{
.predict_knn(knn, ref.cells, cl)
}
\arguments{
\item{knn}{KNN matrix}

\item{ref.cells}{reference cell names}

\item{cl}{cluster assignments}
}
\value{
list with pred.df and pred.prob
}
\description{
Predict cluster membership using KNN
}
\keyword{internal}
