% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constellation.R
\name{.prepare_nodes}
\alias{.prepare_nodes}
\title{Prepare nodes for plotting}
\usage{
.prepare_nodes(cl.center.df, knn.cl.df, node.label, max_size, node_trans)
}
\description{
Prepare nodes for plotting
}
\keyword{internal}
